JHIYA 9/requirements.txt
requests>=2.31.0
dnspython>=2.4.0
python-whois>=0.8.0
ipaddress>=1.0.23
